csproj-12 rename MaybeManaged, perhaps something like Ownership or Managed - renaming solution/namespace to TSMoreland.Extensions.Decorators as the types within are decorators around around generics
