+++ TSMoreland.Extensions.Decorators.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.32804.182
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D44E7720-70CC-4AE3-8FEB-AC8EEFCAAFDF}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSMoreland.Extensions.Decorators.Tests", "test\Decorators.Tests\TSMoreland.Extensions.Decorators.Tests.csproj", "{E7FE64F7-C187-4A11-8F6E-D47822E01D0E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSMoreland.Extensions.Decorators", "src\Decorators\TSMoreland.Extensions.Decorators.csproj", "{E7145F07-90A8-40AE-931E-9CA46EB3E2C4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E7FE64F7-C187-4A11-8F6E-D47822E01D0E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7FE64F7-C187-4A11-8F6E-D47822E01D0E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7FE64F7-C187-4A11-8F6E-D47822E01D0E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7FE64F7-C187-4A11-8F6E-D47822E01D0E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E7145F07-90A8-40AE-931E-9CA46EB3E2C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7145F07-90A8-40AE-931E-9CA46EB3E2C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7145F07-90A8-40AE-931E-9CA46EB3E2C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7145F07-90A8-40AE-931E-9CA46EB3E2C4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {093E2107-91C1-4356-97DB-63104031F37D}
+	EndGlobalSection
+EndGlobal
