--- TSMoreland.MaybeManaged.sln
@@ -1,36 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.32804.182
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSMoreland.MaybeManaged", "src\MaybeManaged\TSMoreland.MaybeManaged.csproj", "{C0DC3A2F-6A0F-43FF-AF0E-FA5E35811EC1}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TSMoreland.MaybeManaged.Tests", "test\MaybeManaged.Tests\TSMoreland.MaybeManaged.Tests.csproj", "{F0FD5E79-14FF-46AE-906B-64D76EA32FBB}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D44E7720-70CC-4AE3-8FEB-AC8EEFCAAFDF}"
-	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C0DC3A2F-6A0F-43FF-AF0E-FA5E35811EC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C0DC3A2F-6A0F-43FF-AF0E-FA5E35811EC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C0DC3A2F-6A0F-43FF-AF0E-FA5E35811EC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C0DC3A2F-6A0F-43FF-AF0E-FA5E35811EC1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F0FD5E79-14FF-46AE-906B-64D76EA32FBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F0FD5E79-14FF-46AE-906B-64D76EA32FBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F0FD5E79-14FF-46AE-906B-64D76EA32FBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F0FD5E79-14FF-46AE-906B-64D76EA32FBB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {093E2107-91C1-4356-97DB-63104031F37D}
-	EndGlobalSection
-EndGlobal
