+++ src/Decorators/IManagedObject.cs
@@ -11,7 +11,7 @@
 // WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
 
-namespace TSMoreland.MaybeManaged;
+namespace TSMoreland.Extensions.Decorators;
 
 /// <summary>
 /// Manages an instance of <see cref="IDisposable"/> ensuring that it is
