--- test/MaybeManaged.Tests/UnitTest1.cs
@@ -1,11 +1,0 @@
-namespace MaybeManaged.Tests
-{
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-
-        }
-    }
-}
